Lots 'o stuff here. See commit for comments.
- Organization (Eclipse packages translate to Github folders) and change
to general naming convention of packages
- This structure illustrates the hierarchy of the code better
- I've added comments to all current classes to explain the structure
- Removed some unused classes (classes that would technically probably
never been used). Piece should suffice as a mediator between Tile and
the unit classes. It will encompass all tiles that can move (which is
only units).

This is just the structure I had on my local machine doing the web app.
Feel free to change anything if you want.